# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/content": {
            "get": {
                "description": "Query content with filters, pagination, and sorting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content"
                ],
                "summary": "List content items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content type filter (e.g. product, post, service)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "published",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword in title and content",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Taxonomy name for filtering (requires term param)",
                        "name": "taxonomy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Taxonomy term slug (requires taxonomy param)",
                        "name": "term",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Meta key for filtering (requires meta_value)",
                        "name": "meta_key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Meta value for filtering (requires meta_key)",
                        "name": "meta_value",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at",
                        "description": "Sort field: created_at, updated_at, published_at, title, sort_order",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "DESC",
                        "description": "Sort direction: ASC or DESC",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page (max 100)",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core_api.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/core_api.contentDTO"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/core_api.pagination"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core_api.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/core_api.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/content/{id}": {
            "get": {
                "description": "Retrieve a single content item by numeric ID or slug (slug requires type-specific endpoint)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content"
                ],
                "summary": "Get content by ID or slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content ID (numeric) or slug (string)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core_api.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/core_api.contentDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core_api.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/core_api.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/core_api.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/core_api.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/types": {
            "get": {
                "description": "Return all registered content type definitions including fields, taxonomies, and rewrite rules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content Types"
                ],
                "summary": "List registered content types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core_api.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core_api.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "core_api.contentDTO": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "published_at": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "core_api.pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "core_api.response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/core_api.apiError"
                },
                "meta": {
                    "$ref": "#/definitions/core_api.pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}